[Chore] #174 - ReviewKeywordView를 관리의 용이를 위해 ReviewView 에서 분리.
